check temp file name before using it

+++ src/upsutl.c
@@ -249,7 +249,7 @@ void upsutl_finish_temp_file( const FILE * const a_stream,
 
   /* we usually tell the file to delete itself.  however the user may
      override this */
-  if (! g_COMPILE_FLAG && ! g_keep_temp_file) {
+  if (g_temp_file_name && ! g_COMPILE_FLAG && ! g_keep_temp_file) {
     fprintf((FILE *)a_stream, "/bin/rm -f %s\n", g_temp_file_name);
   }
 }
